Added the dynamic creation of the CELOE input form.
Added support for dynamically displaying the types of Reasoners.

Cleaned up external repository references as we should be able to go through the AKSW repositories for all of our dependencies.  This is because we now have it set up to proxy to third party repos - this way it can cache what it retrieves and we won't be dependent on third party repos.

A good starting point for the rest of the components - need to move the inputs behind an IConfiguration object.

Added templates for management of different components